Update plugin type and version, add comment to script string
This commit updates the plugin.json file by reducing the plugin types to only the 'helper' type, in accordance with the current functionalities of the plugin. The version has also been bumped up to 1.0.2. Furthermore, a comment line has been added to the __init__.py file for setting the value of the scriptString at the respective address. This provides clarity for code maintainibility and understanding.